+++ src/features/Dashboard/Overview/subscreens/AccountOverviewSubscreen.tsx
@@ -2,10 +2,11 @@ import { Button } from "flowbite-react";
 
 import { AccountDialog } from "@/features/Accounts/AccountDialog";
 import { Account } from "@/features/Accounts/Accounts";
-import { MonthAccordionRecords } from "@/features/Records/CurrentMonthAccordionRecords";
+import { MonthAccordionRecords } from "@/features/Records/Accordions/CurrentMonthAccordionRecords";
 import { useAccountModal } from "@/hooks/useAccountModal";
 import { useDashboard } from "@/shared/hooks/useDashboard";
 import { useDashboardStore } from "@/zustand/provider/dashboard-store-provider";
+import { LastMonthAccordion } from "@/features/Records/Accordions/LastMonthAccordion";
 
 /**
  * This subscreen shows the overview of the account with it's information and records
@@ -47,6 +48,7 @@ export const AccountOverviewSubscreen = () => {
         </Button>
       ) }
       <MonthAccordionRecords records={records} title="Este mes" />
+      <LastMonthAccordion />
     </section>
   )
 }