+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.29.0 (2025-08-01)
+
+### Pull Requests
+[#51](https://github.com/RafaelMoro/next-budget-master/pull/51) | Add last month accordion
+
+
+## v0.28.2 (2025-07-27)
+
+### Pull Requests
+[#50](https://github.com/RafaelMoro/next-budget-master/pull/50) | Add Account Skeleton and fix CLS loading overview screen
+
+
 ## v0.28.1 (2025-07-27)
 
 ### Pull Requests